Move typ face settings related code into separate component
